requirements.txt
bokeh==0.12.14
click==6.7
cycler==0.10.0
Flask==0.12.2
Flask-HTTPAuth==3.2.3
freeport==0.1.9
gunicorn==19.8.1
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.0
numpy==1.14.1
packaging==17.1
pandas==0.22.0
Pillow==5.0.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
tornado==5.0
Werkzeug==0.14.1
WTForms==2.1
